refactor([issue-2532]): extract Universe Builder panels
